project1/pyproject.toml
[tool.poetry]
name = "project1"
version = "0.1.0"
description = ""
authors = ["jahoroshi4y <jahoroshi4y@gmail.com>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.3.0"
appdirs = "1.4.4"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
asgiref = "3.8.1"
asttokens = "2.4.1"
async-lru = "2.0.4"
attrs = "23.2.0"
babel = "2.14.0"
beautifulsoup4 = "4.12.3"
bleach = "6.1.0"
bs4 = "0.0.2"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
comm = "0.2.1"
cssselect = "1.2.0"
debugpy = "1.8.0"
decorator = "5.1.1"
defusedxml = "0.7.1"
distro = "1.9.0"
django = "5.0.4"
django-autoslug = "1.9.9"
djangorestframework = "3.15.1"
executing = "2.0.1"
fake-useragent = "1.5.1"
fastjsonschema = "2.19.1"
fqdn = "1.5.1"
frozenlist = "1.4.1"
gtts = "2.5.1"
h11 = "0.14.0"
httpcore = "1.0.4"
httpx = "0.27.0"
humanize = "4.9.0"
idna = "3.6"
importlib-metadata = "7.0.2"
ipykernel = "6.29.0"
ipython = "8.20.0"
ipywidgets = "8.1.2"
isoduration = "20.11.0"
jedi = "0.19.1"
jinja2 = "3.1.3"
json5 = "0.9.25"
jsonpointer = "2.4"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
jupyter = "1.0.0"
jupyter-console = "6.6.3"
jupyter-events = "0.10.0"
jupyter-lsp = "2.2.5"
jupyter-client = "8.6.0"
jupyter-core = "5.7.1"
jupyter-server = "2.14.0"
jupyter-server-terminals = "0.5.3"
jupyterlab = "4.1.6"
jupyterlab-pygments = "0.3.0"
jupyterlab-server = "2.26.0"
jupyterlab-widgets = "3.0.10"
lxml = "5.1.0"
markupsafe = "2.1.5"
matplotlib-inline = "0.1.6"
mistune = "3.0.2"
multidict = "6.0.5"
nbclient = "0.10.0"
nbconvert = "7.16.3"
nbformat = "5.10.4"
nest-asyncio = "1.6.0"
notebook = "7.1.2"
notebook-shim = "0.2.4"
openai = "0.28.0"
outcome = "1.3.0.post0"
overrides = "7.7.0"
packaging = "23.2"
pandocfilters = "1.5.1"
parse = "1.20.1"
parso = "0.8.3"
pexpect = "4.9.0"
pillow = "10.3.0"
platformdirs = "4.1.0"
prometheus-client = "0.20.0"
prompt-toolkit = "3.0.43"
psutil = "5.9.8"
psycopg2 = "2.9.9"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pycparser = "2.22"
pydantic = "2.7.3"
pydantic-core = "2.18.4"
pyee = "11.1.0"
pyexecjs = "1.5.1"
pygments = "2.17.2"
pyppeteer = "2.0.0"
pyquery = "2.0.0"
pysocks = "1.7.1"
python-dateutil = "2.8.2"
python-json-logger = "2.0.7"
pyyaml = "6.0.1"
pyzmq = "25.1.2"
qtconsole = "5.5.1"
qtpy = "2.4.1"
referencing = "0.34.0"
requests = "2.31.0"
requests-html = "0.10.0"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rpds-py = "0.18.0"
selenium = "4.18.1"
send2trash = "1.8.3"
six = "1.16.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
soupsieve = "2.5"
speedtest = "0.0.1"
speedtest-cli = "2.1.3"
sqlparse = "0.5.0"
stack-data = "0.6.3"
supermemo2 = "2.0.0"
terminado = "0.18.1"
tinycss2 = "1.2.1"
tornado = "6.4"
tqdm = "4.66.2"
traitlets = "5.14.1"
transliterate = "1.10.2"
trio = "0.25.0"
trio-websocket = "0.11.1"
types-python-dateutil = "2.9.0.20240316"
typing-extensions = "4.10.0"
uri-template = "1.3.0"
urllib3 = "1.26.18"
urlopen = "1.0.0"
w3lib = "2.1.2"
wcwidth = "0.2.13"
webcolors = "1.13"
webencodings = "0.5.1"
websocket-client = "1.7.0"
websockets = "10.4"
widgetsnbextension = "4.0.10"
wsproto = "1.2.0"
yarl = "1.9.4"
zipp = "3.18.1"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
